+++ src/pytwitcherapi/chat/__init__.py
@@ -1,6 +1,11 @@
-"""IRC client for interacting with the chat of a channel."""
+"""Package for interacting with the IRC chat of a channel.
+
+The main client for connecting to the channel is :class:`IRCClient`.
+"""
 from __future__ import absolute_import
 
 from .client import *
+from .message import *
+from .connection import *
 
-__all__ = client.__all__
+__all__ = ['IRCClient']