+++ src/pytwitcherapi/chat/message.py
@@ -3,6 +3,9 @@ import re
 import irc.client
 
 
+__all__ = ['Tag', 'Emote', 'Chatter', 'Message3']
+
+
 class Tag(object):
     """An irc v3 tag
 
@@ -258,6 +261,19 @@ class Message3(Message):
         Other user types are mod, global_mod, staff, admin.
         """
         self.set_tags(tags)
+
+    @classmethod
+    def from_event(cls, event):
+        """Create a message from an event
+
+        :param event: the event that was received of type ``pubmsg`` or ``privmsg``
+        :type event: :class:`Event3`
+        :returns: a message that resembles the event
+        :rtype: :class:`Message3`
+        :raises: None
+        """
+        source = Chatter(event.source)
+        return cls(source, event.target, event.arguments[0], event.tags)
 
     def __eq__(self, other):
         """Return True if source, target, text and tags is the same