+++ test/test_chat_client.py
@@ -1,13 +1,13 @@
+import select
 import sys
 import threading
-import time
 
 import irc.server
 import mock
 import pytest
 
 from pytwitcherapi import chat, exceptions
-from pytwitcherapi.chat import message, connection
+from pytwitcherapi.chat import connection
 
 if sys.version_info[0] == 2:
     import Queue as queue
@@ -18,18 +18,37 @@ else:
 class IRCChatClient(chat.IRCClient):
     """Little client that displays, when he joined a channel,
     so we can wait until sending messages"""
-    def __init__(self, session, channel, queuesize=0):
+    def __init__(self, session, channel, queuesize=0, signalat=0):
         super(IRCChatClient, self).__init__(session, channel, queuesize)
-        self.mutex = threading.RLock()
-        self.joined = False
-        self.messagecount = 0  # total messages received
+        self.joined = threading.Event()
+        self.joined_nicks = set()
+        self.signalat = signalat
+        self.messagecount = 0  # total received messaged
+        self.received = threading.Event()
 
     def on_join(self, connection, event):
-        self.joined = True
+        self.joined_nicks.add(event.source.nick)
+        if self.in_connection.nickname in self.joined_nicks and\
+           self.out_connection.nickname in self.joined_nicks:
+            self.joined.set()
 
-    def store_message(self, connection, event):
-        super(IRCChatClient, self).store_message(connection, event)
+    def _connect(self, connection, ip, port, nickname, password):
+        # we have to fake her
+        # the test server cannot handle 2 connections from the same
+        # client
+        # so we use 2 different nicknames for in and out connections
+        if connection is self.in_connection:
+            nickname = nickname + 'in'
+        else:
+            nickname = nickname + 'out'
+        super(IRCChatClient, self)._connect(connection, ip, port,
+                                            nickname, password)
+
+    def store_message(self, *args, **kwargs):
+        super(IRCChatClient, self).store_message(*args, **kwargs)
         self.messagecount += 1
+        if self.messagecount == self.signalat:
+            self.received.set()
 
 
 class IRCServerClient(irc.server.IRCClient):
@@ -39,35 +58,55 @@ class IRCServerClient(irc.server.IRCClient):
     Remember to reset him after each use, by setting the class
     variables to their default.
     """
-    password = None
-    joined = None
-    quited = None
-    messages = []
+    password = queue.Queue()
+    joined = queue.Queue()
+    quited = queue.Queue()
+    messages = queue.Queue()
 
-    def __init__(self, *args, **kwargs):
-        irc.server.IRCClient.__init__(self, *args, **kwargs)
+    def _handle_one(self):
+        """
+        Handle one read/write cycle.
+
+        We override this one to handle read before write
+        """
+        ready_to_read, ready_to_write, in_error = select.select(
+            [self.request], [self.request], [self.request], 0.1)
+
+        if in_error:
+            raise self.Disconnect()
+
+        # See if the client has any commands for us.
+        if ready_to_read:
+            self._handle_incoming()
+
+        # Write any commands to the client
+        while self.send_queue and ready_to_write:
+            msg = self.send_queue.pop(0)
+            try:
+                self._send(msg)
+            except OSError:
+                pass
 
     def handle_quit(self, params):
-        IRCServerClient.quited = (self.client_ident(), params)
+        IRCServerClient.quited.put((self.client_ident(), params))
         return irc.server.IRCClient.handle_quit(self, params)
 
     def handle_join(self, params):
-        IRCServerClient.joined = (self.client_ident(), params)
+        IRCServerClient.joined.put((self.client_ident(), params))
         return irc.server.IRCClient.handle_join(self, params)
 
     def handle_pass(self, params):
-        IRCServerClient.password = (self.client_ident(), params)
+        IRCServerClient.password.put((self.client_ident(), params))
 
     def handle_privmsg(self, params):
-        IRCServerClient.messages.append((self.client_ident(), params))
+        IRCServerClient.messages.put((self.client_ident(), params))
         return irc.server.IRCClient.handle_privmsg(self, params)
-
 
 @pytest.fixture(scope='function')
 def mock_get_waittime(monkeypatch):
     m = mock.Mock()
     m.return_value = 0
-    monkeypatch.setattr(connection.ServerConnection3, 'get_waittime', m)
+    monkeypatch.setattr(chat.ServerConnection3, 'get_waittime', m)
 
 
 @pytest.fixture(scope='function')
@@ -75,10 +114,10 @@ def ircserver(request):
     ircserver = irc.server.IRCServer(('0.0.0.0', 0), IRCServerClient)
 
     def fin():
-        IRCServerClient.password = None
-        IRCServerClient.joined = None
-        IRCServerClient.quited = None
-        IRCServerClient.messages = []
+        IRCServerClient.password = queue.Queue()
+        IRCServerClient.joined = queue.Queue()
+        IRCServerClient.quited = queue.Queue()
+        IRCServerClient.messages = queue.Queue()
     request.addfinalizer(fin)
     return ircserver
 
@@ -143,40 +182,61 @@ def ircthreads(request, ircserver, ircclient):
     t2.start()
 
 
+def wait_for_client_joined(client, timout=1.0):
+    client.joined.wait(1.0)  # wait for in and out connection
+    assert client.joined.is_set(), 'Both connections (in and out) of client %s did not join.' % client
+
+
 def simulate_client_server_interaction(ircserver, ircclient):
     """Wait for the client to join, then send messages, then quit."""
-    while not ircclient.joined:
-        time.sleep(0)
+    wait_for_client_joined(ircclient)
 
     ircclient.send_msg('hahaha')
     ircclient.send_msg('hihihi')
+
     # wait till messages have arrived at the server
-    while not len(IRCServerClient.messages) == 2:
-        time.sleep(0)
+    messages = set()
+    for i in range(2):
+        messages.add(IRCServerClient.messages.get(timeout=1))
 
     ircclient.shutdown()
-    # wait till the clien (in other thread) is actually shut down
-    while not IRCServerClient.quited:
-        time.sleep(0)
+    # wait till the client (in other thread) is actually shut down
+    # wait for in and out connection
+    quited = set()
+    for i in range(2):
+        quited.add(IRCServerClient.quited.get(timeout=1))
+    return messages, quited
 
 
-@pytest.mark.timeout(10)
 def test_client(ircserver, ircclient, ircthreads, access_token):
-    simulate_client_server_interaction(ircserver, ircclient)
-    testuser = 'testuser!testuser@localhost'
-    expectedpw = ('None!None@localhost', 'oauth:%s' % access_token)
-    expectedjoin = (testuser, '#test_channel')
-    expectedmessages = [(testuser, '#test_channel :hahaha'),
-                        (testuser, '#test_channel :hihihi')]
-    assert IRCServerClient.password == expectedpw
-    assert IRCServerClient.joined == expectedjoin
-    assert IRCServerClient.quited == (testuser, '')
-    assert IRCServerClient.messages == expectedmessages
+    messages, quited = simulate_client_server_interaction(ircserver, ircclient)
+    testchannel = '#test_channel'
+    testuser = 'testuser%s!testuser%s@localhost'
+    testuserin = testuser % ('in', 'in')
+    testuserout = testuser % ('out', 'out')
+    expectedpw = (('None!None@localhost', 'oauth:%s' % access_token),
+                  ('None!None@localhost', 'oauth:%s' % access_token))
+    expectedjoin = ((testuserin, testchannel),
+                    (testuserout, testchannel))
+    expectedmessages = ((testuserout, '#test_channel :hahaha'),
+                        (testuserout, '#test_channel :hihihi'))
+    passwords = set()
+    joined = set()
+    for i in range(2):
+        passwords.add(IRCServerClient.password.get(timeout=1))
+        joined.add(IRCServerClient.joined.get(timeout=1))
+
+    assert passwords == set(expectedpw)
+    assert joined == set(expectedjoin)
+    assert quited == set(((testuserin, ''),
+                          (testuserout, '')))
+    assert messages == set(expectedmessages)
 
 
 def test_disconnect(ircserver, ircclient):
     ircclient.channel = None
-    assert not ircclient.connection.connected, 'The client should disconnect the connection, \
+    for c in [ircclient.in_connection, ircclient.out_connection]:
+        assert not c.connected, 'The client should disconnect the connection, \
 when you set the channel to None'
 
 
@@ -197,14 +257,13 @@ def assert_client_got_message(client, message):
         raise AssertionError('ircclient did not store the message in the message queue')
 
 
-@pytest.mark.timeout(10)
 def test_message_queue(ircclient, ircclient2, ircthreads, ircclient2thread):
-    c = message.Chatter('testuser2!testuser2@localhost')
-    m1 = message.Message3(c, '#test_channel', 'mic check')
-    m2 = message.Message3(c, '#test_channel', 'onetwo')
+    c = chat.Chatter('testuser2out!testuser2out@localhost')
+    m1 = chat.Message3(c, '#test_channel', 'mic check')
+    m2 = chat.Message3(c, '#test_channel', 'onetwo')
 
-    while not ircclient2.joined:
-        time.sleep(0)
+    wait_for_client_joined(ircclient)
+    wait_for_client_joined(ircclient2)
 
     for m in [m1, m2]:
         ircclient2.send_msg(m.text)
@@ -217,21 +276,21 @@ def _send_messages(client):
     """
     # send 7 private
     for i in range(7):
-        client.privmsg('testuser2', '%i' % i)
+        client.privmsg('testuser2in', '%i' % i)
     # the rest public
     for i in range(7, 15):
         client.privmsg('#test_channel', '%i' % i)
 
 
-@pytest.mark.timeout(10)  # if it times out, maybe the store_message method uses a blocking put
 def test_message_queue_full(ircclient, ircclient2, ircthreads, ircclient2thread):
-    while not ircclient2.joined:
-        time.sleep(0)
+    ircclient2.signalat = 15
+    wait_for_client_joined(ircclient)
+    wait_for_client_joined(ircclient2)
 
     _send_messages(ircclient)
 
-    while ircclient2.messagecount != 15:
-        time.sleep(0)
+    ircclient2.received.wait(5)
+    assert ircclient2.received.is_set(), 'Did not receive all messages'
 
     assert ircclient2.messages.full(), '15 messages have been sent. The queue should be full.'
     assert ircclient2.messages.qsize() == 10, '15 messages have been sent, but the queue should have a maxsize of 10'