+++ test/test_chat_message.py
@@ -1,12 +1,13 @@
 import pytest
 
+from pytwitcherapi import chat
 from pytwitcherapi.chat import message
 
 
 @pytest.mark.parametrize('text,expected', [('hallo!', 'hallo!'),
                                            ('ciao bella', 'ciao ...')])
 def test_repr(text, expected):
-    source = message.Chatter('me')
+    source = chat.Chatter('me')
     target = '#you'
     msg = message.Message(source=source, target=target, text=text)
     assert repr(msg) == '<Message me to #you: %s>' % expected
@@ -25,17 +26,17 @@ def assert_message_attrs(message, **kwargs):
         assert messagevalue == expected
 
 
-TAGS1 = [message.Tag('color', '#0000FF'),
-         message.Tag('emotes', '36031:0-7,22-30/40894:9-18'),
-         message.Tag('subscriber', '1'),
-         message.Tag('turbo', '1'),
-         message.Tag('user-type', 'mod')]
+TAGS1 = [chat.Tag('color', '#0000FF'),
+         chat.Tag('emotes', '36031:0-7,22-30/40894:9-18'),
+         chat.Tag('subscriber', '1'),
+         chat.Tag('turbo', '1'),
+         chat.Tag('user-type', 'mod')]
 
-TAGS2 = [message.Tag('display-name', 'haha'),
-         message.Tag('emotes', None)]
+TAGS2 = [chat.Tag('display-name', 'haha'),
+         chat.Tag('emotes', None)]
 
-EMOTES1 = [message.Emote(36031, [(0, 7), (22, 30)]),
-           message.Emote(40894, [(9, 18)])]
+EMOTES1 = [chat.Emote(36031, [(0, 7), (22, 30)]),
+           chat.Emote(40894, [(9, 18)])]
 
 ATTRS1 = {'color': '#0000FF',
           'emotes': EMOTES1,
@@ -53,5 +54,5 @@ ATTRS2 = {'color': None,
 @pytest.mark.parametrize('tags,attrs', [(TAGS1, ATTRS1),
                                         (TAGS2, ATTRS2)])
 def test_set_tags(tags, attrs):
-    m = message.Message3('', '', '', tags)
+    m = chat.Message3('', '', '', tags)
     assert_message_attrs(m, **attrs)